// src/lib/api/models/PropertyTypeEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `apartment` - Apartment
 * * `house` - House
 * * `villa` - Villa
 * * `condo` - Condo
 * * `townhouse` - Townhouse
 * * `studio` - Studio
 * * `cabin` - Cabin
 * * `cottage` - Cottage
 * * `other` - Other
 */
export enum PropertyTypeEnum {
    APARTMENT = 'apartment',
    HOUSE = 'house',
    VILLA = 'villa',
    CONDO = 'condo',
    TOWNHOUSE = 'townhouse',
    STUDIO = 'studio',
    CABIN = 'cabin',
    COTTAGE = 'cottage',
    OTHER = 'other',
}